Project/requirements.txt
# Core
SpeechRecognition>=3.8.1
pygame>=2.1.3
gTTS>=2.3.0
pyttsx3>=2.90
requests>=2.31.0
python-dotenv>=1.0.0
pyautogui>=0.9.53
pillow>=9.5.0

# Optional (better wakeword)
pvporcupine>=2.2.0; platform_system!="Linux" or platform_machine!="armv7l"
# If using spaCy for advanced NER/intent (optional)
spacy>=3.5.3

# OpenAI python client (newer clients)
openai>=0.27.8
